# lms-backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: MOOC Learning Platform API
  description: >
    Đặc tả API cho nền tảng học tập trực tuyến theo mô hình MOOC.
    Bao gồm 3 nhóm tác nhân: Student (Học viên), Instructor (Giảng viên), Admin (Quản trị viên).
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local dev server
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Đăng ký tài khoản mới (mặc định role Student)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password, fullName]
              properties:
                username: { type: string, example: anhlinh23 }
                email: { type: string, format: email }
                password: { type: string, format: password, minLength: 6 }
                fullName: { type: string }
      responses:
        '201':
          description: Đăng ký thành công
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '400':
          description: Username hoặc email đã tồn tại
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Đăng nhập, trả về access token và refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Đăng nhập thành công
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Sai email hoặc mật khẩu
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /categories:
    get:
      tags: [Category]
      summary: Danh sách danh mục khóa học (công khai)
      security: []
      responses:
        '200':
          description: Danh sách danh mục
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CategoryResponse' }
    post:
      tags: [Category]
      summary: Tạo danh mục mới (chỉ Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string }
                slug: { type: string }
                description: { type: string }
      responses:
        '201':
          description: Tạo thành công
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryResponse' }
  /categories/{id}:
    put:
      tags: [Category]
      summary: Cập nhật danh mục (chỉ Admin)
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                slug: { type: string }
                description: { type: string }
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryResponse' }
    delete:
      tags: [Category]
      summary: Xóa danh mục (chỉ Admin)
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204':
          description: Xóa thành công

  # ==========================================================
  # COURSE - PUBLIC (Student / khách)
  # ==========================================================
  /courses:
    get:
      tags: [Course - Public]
      summary: Tìm kiếm, lọc, phân trang khóa học đã xuất bản
      security: []
      parameters:
        - name: category
          in: query
          schema: { type: string }
          description: Lọc theo slug danh mục
        - name: keyword
          in: query
          schema: { type: string }
          description: Tìm theo tên khóa học
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 12 }
      responses:
        '200':
          description: Danh sách khóa học (phân trang)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoursePageResponse' }
  /courses/{id}:
    get:
      tags: [Course - Public]
      summary: Xem chi tiết khóa học (bao gồm danh sách chương/bài)
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Chi tiết khóa học
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CourseDetailResponse' }
        '404':
          description: Không tìm thấy khóa học

  # ==========================================================
  # COURSE - INSTRUCTOR (quản lý khóa học, chương, bài học)
  # ==========================================================
...
  /instructor/courses/{id}:
    put:
      tags: [Course - Instructor]
      summary: Cập nhật thông tin khóa học (chỉ chủ sở hữu)
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CourseCreateRequest' }
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CourseResponse' }
    delete:
      tags: [Course - Instructor]
      summary: Xóa khóa học (chỉ khi ở trạng thái Draft)
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204':
          description: Xóa thành công
...
  /lessons/{id}:
    put:
      tags: [Course - Instructor]
      summary: Cập nhật bài học
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LessonCreateRequest' }
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LessonResponse' }
    delete:
      tags: [Course - Instructor]
      summary: Xóa bài học
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204':
          description: Xóa thành công

  # ==========================================================
  # COURSE - ADMIN (kiểm duyệt)
  # ==========================================================
...
  /admin/courses/{id}/approve:
    put:
      tags: [Course - Admin]
      summary: Duyệt và xuất bản khóa học (Pending → Published)
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Duyệt thành công
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CourseResponse' }
  /admin/courses/{id}/reject:
    put:
      tags: [Course - Admin]
      summary: Từ chối khóa học (Pending → Rejected)
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, description: Lý do từ chối }
      responses:
        '200':
          description: Từ chối thành công
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CourseResponse' }

  # ==========================================================
  # ENROLLMENT - Student
  # ==========================================================
...
  /admin/users:
    get:
      tags: [Admin - User Management]
      summary: Danh sách tất cả người dùng (lọc theo role, trạng thái)
      parameters:
        - name: role
          in: query
          schema: { type: string, enum: [Admin, Instructor, Student] }
        - name: keyword
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Danh sách người dùng
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UserResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        message: { type: string }
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        user: { $ref: '#/components/schemas/UserResponse' }
    UserResponse:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string }
        fullName: { type: string }
        avatarUrl: { type: string }
        role: { type: string, enum: [Admin, Instructor, Student] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    CategoryResponse:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
    CourseCreateRequest:
      type: object
      required: [title, categoryId]
      properties:
        title: { type: string }
        description: { type: string }
        thumbnailUrl: { type: string }
        categoryId: { type: integer }
    CourseResponse:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        thumbnailUrl: { type: string }
        categoryName: { type: string }
        instructorName: { type: string }
        status: { type: string, enum: [Draft, Pending, Published, Rejected] }
        createdAt: { type: string, format: date-time }
    CoursePageResponse:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/CourseResponse' }
        page: { type: integer }
        totalPages: { type: integer }
        totalElements: { type: integer }
    CourseDetailResponse:
      allOf:
        - $ref: '#/components/schemas/CourseResponse'
        - type: object
          properties:
            sections:
              type: array
              items: { $ref: '#/components/schemas/SectionResponse' }
    SectionResponse:
      type: object
      properties:
        id: { type: integer }
        courseId: { type: integer }
        title: { type: string }
        orderIndex: { type: integer }
        lessons:
          type: array
          items: { $ref: '#/components/schemas/LessonResponse' }
    LessonCreateRequest:
      type: object
      required: [title, contentType]
      properties:
        title: { type: string }
        contentType: { type: string, enum: [Video, Document, Text] }
        contentUrl: { type: string }
        durationMinutes: { type: integer, default: 0 }
        orderIndex: { type: integer }
    LessonResponse:
      type: object
      properties:
        id: { type: integer }
        sectionId: { type: integer }
        title: { type: string }
        contentType: { type: string, enum: [Video, Document, Text] }
        contentUrl: { type: string }
        durationMinutes: { type: integer }
        orderIndex: { type: integer }
...